Create App.js
Here, In the new QueryClient, we have some default options like queries and mutation. In queries, we do have a property named refetchOnWindowsFocus. if refetchOnWindowsFocus is true, then every time the data will be fetched from the API even if you change the tab, and if refetchOnWindowsFocus  is false, then the existing data will be there no matter how many tabs you change